Move all playlist info to seperate file
