+++ src/Components/StatsBox/StatsBox.css
@@ -1,0 +1,10 @@
+.totals {
+  text-align: center;
+  font-family: sans-serif;
+  font-weight: 100;
+  font-size: 40px;
+  color: #0922e4;
+  text-shadow: 0 0 4px #0922e4;
+  margin: 0px;
+  margin-top: 15px;
+}
