+++ src/Components/ViewSubtitle/ViewSubtitle.css
@@ -1,0 +1,10 @@
+.small-header {
+  text-align: center;
+  font-family: sans-serif;
+  font-weight: 100;
+  font-size: 26px;
+  color: #857e7a;
+  text-shadow: 0 0 4px #857e7a;
+  margin: 0px;
+  margin-top: 10px;
+}
