+++ src/Components/ViewSubtitle/ViewSubtitle.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './ViewSubtitle.css'
+
+const ViewSubtitle = ({id}) => {
+  return (
+    <h2 className='small-header'>If you feel {id} today...<br></br>You're not alone. There are:</h2>
+  )
+}
+
+export default ViewSubtitle;
